Add examples to illustrate the Shell grammar
I defined separate scripts for non-terminal symbols of the grammar,
and tried to mimic the grammar rules to construct examples of the
syntax at each level.

This is a work in progress, covering the three first rules of the
Shell grammar so far.

Note that to simplify the examples, I assigned '.' to the PATH
to look for the commands unprefixed in the current folder, which
requires to call these scripts directly from the parent folder.