Remove line of multiple examples of compound_list to avoid confusion
In other scripts so far, I have separated examples by a newline.
In this script, several examples were grouped on the same line
to indicate that no newline was required between them. I removed
this example to limit induced confusion.

While the list is defined in the grammar as a single term,
it is the rule for the term that actually introduces the
recursion which allows to have a list of terms in a term.